RED-377 move helpers to root utils
