+++ packages/redshift-types/package.json
@@ -11,7 +11,8 @@
   "dependencies": {
     "@types/bip32": "^1.0.1",
     "@types/bitcoinjs-lib": "4.0.0",
-    "@types/web3": "^1.0.18"
+    "@types/web3": "^1.0.18",
+    "bignumber.js": "^8.0.2"
   },
   "devDependencies": {}
 }
