+++ packages/redshift-utils/src/numeric.ts
