+++ packages/redshift-utils/src/rpc/ethereum/ethereum-rpc-client.ts
@@ -1,7 +1,7 @@
 import { Network, RpcConnectionConfig, Subnet } from '@radar/redshift-types';
 import BigNumber from 'bignumber.js';
 import { Tx } from 'web3/eth/types';
-import { hexToBigNumber } from '../../helpers/numeric';
+import { hexToBigNumber } from '../../numeric';
 import { BaseRpcClient } from '../base-rpc-client';
 
 export class EthereumRpcClient extends BaseRpcClient {
